refactor(dsl-protobuf): split protobuf DSL scope contracts
