+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/EnumValueScope.kt
@@ -1,0 +1,6 @@
+package me.liam.microsmith.dsl.schemas.protobuf
+
+import me.liam.microsmith.dsl.core.MicrosmithDsl
+
+@MicrosmithDsl
+interface EnumValueScope : FieldScope
