+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/FieldScope.kt
@@ -1,0 +1,5 @@
+package me.liam.microsmith.dsl.schemas.protobuf
+
+interface FieldScope {
+    fun index(index: Int)
+}
