+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/ReferenceFieldScope.kt
@@ -1,0 +1,8 @@
+package me.liam.microsmith.dsl.schemas.protobuf
+
+import me.liam.microsmith.dsl.core.MicrosmithDsl
+
+@MicrosmithDsl
+interface ReferenceFieldScope :
+    OneofReferenceFieldScope,
+    ScalarFieldScope
