+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/ScalarFieldScope.kt
@@ -1,0 +1,10 @@
+package me.liam.microsmith.dsl.schemas.protobuf
+
+import me.liam.microsmith.dsl.core.MicrosmithDsl
+
+@MicrosmithDsl
+interface ScalarFieldScope : FieldScope {
+    fun optional()
+
+    fun repeated()
+}
